+++ ContentPatcher/release-notes.md
@@ -10,7 +10,8 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 -->
 ## Upcoming release
-* Content packs can now [add custom locations](docs/author-guide.md#custom-locations), and the `patch summary` command now includes custom locations.
+* Added support for [creating custom locations](docs/author-guide.md#custom-locations).
+* Added `AddWarps` field for [`EditMap` patches](https://github.com/Pathoschild/StardewMods/blob/develop/ContentPatcher/docs/author-guide.md#editmap).
 
 ## 1.20
 Released 06 February 2021. See the [release highlights](https://www.patreon.com/posts/47213526).